% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IARIMAXoid_Pro.R
\name{IARIMAXoid_Pro}
\alias{IARIMAXoid_Pro}
\title{Run I-ARIMAX algorithm.}
\usage{
IARIMAXoid_Pro(
  dataframe,
  min_n_subject = 20,
  minvar = 0.01,
  y_series,
  x_series,
  id_var,
  metaanalysis = TRUE
)
}
\arguments{
\item{dataframe}{Your dataframe.}

\item{min_n_subject}{The minimum number of non NA cases to run the analyses. It will filter cases with more NA's than the threshold. Defaults to 20.}

\item{minvar}{The minimum variance for both series (&) to include a case. Defaults to 0.01.}

\item{y_series}{A string containing the name of your dependent variable y.}

\item{x_series}{A string containing the name of your independent variable x.}

\item{id_var}{A string containing your id variable.}

\item{metaanalysis}{Bool to run a random effects meta-analysis or not.}
}
\value{
A list containing a dataframe with the ARIMA parameteres, plus the xreg parameter (the beta value for your x_series) together with their std.errors. If metaanalysis = TRUE, will also output a random effects meta analysis.
}
\description{
Run I-ARIMAX algorithm.
}
