% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IARIMAXoid_Pro_Lag.R
\name{IARIMAXoid_Pro_Lag}
\alias{IARIMAXoid_Pro_Lag}
\title{Run dynamic modelling I-ARIMAX algorithm.}
\usage{
IARIMAXoid_Pro_Lag(
  dataframe,
  min_n_subject = 20,
  minvar = 0.01,
  y_series,
  x_series,
  id_var,
  metaanalysis = TRUE,
  hlm_compare = FALSE,
  timevar = NULL,
  weight_rma = FALSE,
  weight_rma_var = NULL,
  correlation_method = "pearson"
)
}
\arguments{
\item{dataframe}{Your dataframe.}

\item{min_n_subject}{The minimum number of non NA cases to run the analyses. It will filter cases with more NA's than the threshold. Defaults to 20.}

\item{minvar}{The minimum variance for both series (&) to include a case. Defaults to 0.01.}

\item{y_series}{A string containing the name of your dependent variable y.}

\item{x_series}{A string containing the name of your independent variable x.}

\item{id_var}{A string containing your id variable.}

\item{metaanalysis}{Bool to run a random effects meta-analysis or not.}

\item{hlm_compare}{Optional, to create a comparison with an HLM model, default is FALSE.}

\item{timevar}{If hlm_compare is TRUE, then a time variable is needed, default is NULL.}

\item{weight_rma}{If adding an exogenous weight variable to the RMA model.}

\item{weight_rma_var}{Select weight RMA variable. Defaults as NULL, if NULL then the number of valid observations for Y AND X will be used (!is.na).}

\item{correlation_method}{Select method for raw semi-partial correlations. Options are: 'spearman', 'pearson' or 'kendall'. Defaults to 'pearson'.}
}
\value{
A list containing a dataframe with the ARIMA parameters, plus the xreg parameter (the beta value for your x_series and x_series lag) together with their std.errors. If metaanalysis = TRUE, will also output a random effects meta analysis. If hlm_compare = TRUE, will also output a model comparison with HLM.
}
\description{
Run dynamic modelling I-ARIMAX algorithm.
}
