% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idionomics-package.R
\docType{package}
\name{idionomics-package}
\alias{idionomics}
\alias{idionomics-package}
\title{idionomics: Conduct Idiographic-Nomothetic Analyses for Psych Science}
\description{
A toolkit for idionomic science, a research philosophy that places the individual at the center of analysis. Rather than fitting a single model to the group and treating between-person averages as the primary finding, idionomic methods model each person's time series independently and then aggregate upward using random-effects meta-analysis that explicitly represents and quantifies between-person heterogeneity. The package targets intensive longitudinal data such as experience-sampling, ecological momentary assessment, and daily diary designs. It provides a complete pipeline: data quality screening ('i_screener()'), within-person standardization ('pmstandardize()'), linear detrending ('i_detrender()'), per-subject ARIMAX modeling and meta-analysis ('iarimax()'), individual p-values ('i_pval()'), Sign Divergence and Equisyncratic Null tests ('sden_test()'), and directed loop detection ('looping_machine()').
}
\references{
Hernandez, C., Sahdra, B. K., Ciarrochi, J., & Hayes, S. C. (2024).
Idionomic Assessment of Mindfulness. \emph{Handbook of Assessment in Mindfulness
Research}, 1--27. \url{https://doi.org/10.1007/978-3-030-77644-2_136-1}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cristobalehc/idionomics}
  \item Report bugs at \url{https://github.com/cristobalehc/idionomics/issues}
}

}
\author{
\strong{Maintainer}: Cristóbal Hernández \email{cristobal.ehc@gmail.com}

Authors:
\itemize{
  \item Joseph Ciarrochi \email{ciarrochij@gmail.com}
  \item Steven Hayes \email{stevenchayes@gmail.com}
  \item Baljinder Sahdra \email{baljinder.sahdra@acu.edu.au}
}

}
