% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputatron_2000.R
\name{imputatron_2000}
\alias{imputatron_2000}
\title{Easy copyMeans imputation for time-series.}
\usage{
imputatron_2000(data, id_col, time_col, cols_of_interest)
}
\arguments{
\item{data}{A string containing your dataframe name.}

\item{id_col}{A string containing your ID column.}

\item{time_col}{A string containing your time column.}

\item{cols_of_interest}{A vector of strings containing the columns you wish to impute.}
}
\value{
It returns a list with a new dataframe with imputed values, longdata objects with and without imputation, your filtered data, wide data used to create the longdata object, and a list with indexes to create the longdata object.
}
\description{
Easy copyMeans imputation for time-series.
}
