% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_detrender.R
\name{i_detrender}
\alias{i_detrender}
\title{Detrend idiographic time series by time variable.}
\usage{
i_detrender(df, cols, idvar, timevar, explanation = TRUE, append = TRUE)
}
\arguments{
\item{df}{A dataframe.}

\item{cols}{A vector of strings with the variable names you would like to detrend.}

\item{idvar}{Your ID variable. It currently accepts only one.}

\item{timevar}{The time variable used to detrend yout data.}

\item{explanation}{Boolean to print explanation of the function (default is TRUE).}

\item{append}{Boolean to append new standardized columns to your original dataframe or not (default is TRUE)}
}
\value{
Will return a dataframe with new detrended columns appended at the end.
}
\description{
Detrend idiographic time series by time variable.
}
