% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_pval.R
\name{i_pval}
\alias{i_pval}
\title{Case by case p-value calculation based on t-distribution.}
\usage{
i_pval(iarimax_object, feature = "xreg")
}
\arguments{
\item{iarimax_object}{An iarimax object.}

\item{feature}{Feature name to calculate p-value. Defaults to "xreg".}
}
\value{
Returns an updated version of results_df within the iarimax_object with p-values for the specific feature.
}
\description{
Case by case p-value calculation based on t-distribution.
}
