% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iarimax.R
\name{iarimax}
\alias{iarimax}
\title{Run I-ARIMAX algorithm.}
\usage{
iarimax(
  dataframe,
  min_n_subject = 20,
  minvar = 0.01,
  y_series,
  x_series,
  focal_predictor = NULL,
  id_var,
  timevar,
  correlation_method = "pearson",
  keep_models = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dataframe}{Your dataframe.}

\item{min_n_subject}{The minimum number of non NA cases to run the analyses. It will filter cases with more NA's than the threshold. Defaults to 20.}

\item{minvar}{The minimum variance for both series (&) to include a case. Defaults to 0.01.}

\item{y_series}{A string containing the name of your dependent variable y.}

\item{x_series}{A character vector containing the name(s) of your predictor variable(s).}

\item{focal_predictor}{A string with the name of the predictor to use in the meta-analysis. Required when x_series has more than one variable. When x_series is a single variable, defaults to that variable.}

\item{id_var}{A string containing your id variable.}

\item{timevar}{Required to arrange timeseries.}

\item{correlation_method}{Select method for raw correlations. Options are: 'spearman', 'pearson' or 'kendall'. Defaults to 'pearson'.}

\item{keep_models}{If TRUE, will keep original arimax models in a list.}

\item{verbose}{If TRUE, prints progress messages during filtering and model fitting. Defaults to FALSE.}
}
\value{
A list containing a dataframe with the ARIMA parameters, plus the xreg parameters (the beta values for your x_series) together with their std.errors, and a random effects meta analysis on the focal predictor.
}
\description{
Run I-ARIMAX algorithm.
}
