% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sden_test.R
\name{sden_test}
\alias{sden_test}
\title{Run Sign Divergence or Equisyncratic Null tests.}
\usage{
sden_test(
  iarimax_object,
  alpha_arimax = 0.05,
  alpha_binom = NULL,
  test = "auto"
)
}
\arguments{
\item{iarimax_object}{Your iarimax object.}

\item{alpha_arimax}{Critical value for arimax parameter significance. Defaults to 0.05.}

\item{alpha_binom}{Critical value for binomial test. Defaults to alpha_arimax for ENT, and alpha_arimax/2 for SDT}

\item{test}{type of test, default to "auto". "auto", "SDT", and "ENT" are supported.}
}
\value{
The appropriate sden test.
}
\description{
Run Sign Divergence or Equisyncratic Null tests.
}
