% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.sden_results}
\alias{summary.sden_results}
\title{Summary method for sden_results objects.}
\usage{
\method{summary}{sden_results}(object, ...)
}
\arguments{
\item{object}{An object of class \code{sden_results}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{object}, called for its side effect of printing.
}
\description{
Summary method for sden_results objects.
}
